Update main.py
Rewrote csv reader module, using basic python csv instead of numpy, due to an error with stars such as BLG 00712 where the stars start with 3 digit dates rather than 4, and so there is a leading space. Now dates and brightnesses are read with csv skipinitialspace, and read and sliced with 2d tuple.