Update main.py
Improve autosaver, fixed issue where RA/dec coordinates were not being read correctly, as the metadata was not formatted correctly with missing columns for some entries. Also implemented workaround for issue in GaussNewton where the inverse of the dot product of J and J transpose was unable to calculated as determinant = 0. Stars where this happens have graphs that don't work, and the names are logged in an error file, matrixbugloc.